@@ -1,10 +1,10 @@
 name: paged_datatable
 description: A brand new way of creating paginated DataTables in Flutter with sorting and filters
-version: 2.1.11
+version: 2.2.0
 homepage: https://github.com/tomasweigenast/paged-datatable
 
 environment:
-  sdk: "^3.0.0"
+  sdk: "^3.8.0"
 
 dependencies:
   flutter:
@@ -15,7 +15,7 @@ dependencies:
 
 dev_dependencies:
   build_runner: ^2.4.8
-  flutter_lints: ^5.0.0
+  flutter_lints: ^6.0.0
   intl_utils: ^2.8.5
 
 flutter:
