@@ -28,31 +28,35 @@ class MessageLookup extends MessageLookupByLibrary {
 
   final messages = _notInlinedMessages(_notInlinedMessages);
   static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
-        "applyFilterButtonText":
-            MessageLookupByLibrary.simpleMessage("Anwenden"),
-        "cancelFilteringButtonText":
-            MessageLookupByLibrary.simpleMessage("Abbrechen"),
-        "editableColumnCancelButtonText":
-            MessageLookupByLibrary.simpleMessage("Abbrechen"),
-        "editableColumnSaveChangesButtonText":
-            MessageLookupByLibrary.simpleMessage("Änderungen speichern"),
-        "filterByTitle": MessageLookupByLibrary.simpleMessage("Filtern nach"),
-        "nextPageButtonText":
-            MessageLookupByLibrary.simpleMessage("Nächste Seite"),
-        "noItemsFoundText":
-            MessageLookupByLibrary.simpleMessage("Keine Elemente gefunden"),
-        "pageIndicatorText": m0,
-        "previousPageButtonText":
-            MessageLookupByLibrary.simpleMessage("Vorherige Seite"),
-        "refreshText": MessageLookupByLibrary.simpleMessage("Aktualisieren"),
-        "refreshedAtText": m1,
-        "removeAllFiltersButtonText":
-            MessageLookupByLibrary.simpleMessage("Entfernen"),
-        "removeFilterButtonText":
-            MessageLookupByLibrary.simpleMessage("Diesen Filter entfernen"),
-        "rowsPerPageText":
-            MessageLookupByLibrary.simpleMessage("Zeilen pro Seite"),
-        "showFilterMenuTooltip": MessageLookupByLibrary.simpleMessage("Filter"),
-        "totalElementsText": m2
-      };
+    "applyFilterButtonText": MessageLookupByLibrary.simpleMessage("Anwenden"),
+    "cancelFilteringButtonText": MessageLookupByLibrary.simpleMessage(
+      "Abbrechen",
+    ),
+    "editableColumnCancelButtonText": MessageLookupByLibrary.simpleMessage(
+      "Abbrechen",
+    ),
+    "editableColumnSaveChangesButtonText": MessageLookupByLibrary.simpleMessage(
+      "Änderungen speichern",
+    ),
+    "filterByTitle": MessageLookupByLibrary.simpleMessage("Filtern nach"),
+    "nextPageButtonText": MessageLookupByLibrary.simpleMessage("Nächste Seite"),
+    "noItemsFoundText": MessageLookupByLibrary.simpleMessage(
+      "Keine Elemente gefunden",
+    ),
+    "pageIndicatorText": m0,
+    "previousPageButtonText": MessageLookupByLibrary.simpleMessage(
+      "Vorherige Seite",
+    ),
+    "refreshText": MessageLookupByLibrary.simpleMessage("Aktualisieren"),
+    "refreshedAtText": m1,
+    "removeAllFiltersButtonText": MessageLookupByLibrary.simpleMessage(
+      "Entfernen",
+    ),
+    "removeFilterButtonText": MessageLookupByLibrary.simpleMessage(
+      "Filter entfernen",
+    ),
+    "rowsPerPageText": MessageLookupByLibrary.simpleMessage("Zeilen pro Seite"),
+    "showFilterMenuTooltip": MessageLookupByLibrary.simpleMessage("Filter"),
+    "totalElementsText": m2,
+  };
 }
