@@ -1,10 +1,10 @@
-library paged_datatable;
-
 export './l10n/generated/l10n.dart';
 export './src/paged_datatable.dart';
 export './src/column_size.dart';
 export './src/column_format.dart';
 export './src/configuration.dart';
+export './src/filter_bar_controller.dart';
 export './src/footer.dart';
 export './src/theme.dart';
-export './src/table_controller_notifier.dart';
+export './src/table_controller_provider.dart';
+export './src/table_state.dart';
