@@ -24,8 +24,7 @@ final class FixedColumnSize extends ColumnSize {
   int get hashCode => size.hashCode;
 
   @override
-  bool operator ==(Object other) =>
-      other is FixedColumnSize && other.size == size;
+  bool operator ==(Object other) => other is FixedColumnSize && other.size == size;
 
   @override
   bool get isFixed => true;
@@ -39,22 +38,20 @@ final class FractionalColumnSize extends ColumnSize {
   final double _fraction;
 
   const FractionalColumnSize(double fraction)
-      : _fraction = fraction,
-        assert(fraction > 0, "Fraction cannot be less than or equal to zero.");
+    : _fraction = fraction,
+      assert(fraction > 0, "Fraction cannot be less than or equal to zero.");
 
   @override
   int get hashCode => fraction.hashCode;
 
   @override
-  bool operator ==(Object other) =>
-      other is FractionalColumnSize && other.fraction == fraction;
+  bool operator ==(Object other) => other is FractionalColumnSize && other.fraction == fraction;
 
   @override
   double get fraction => _fraction;
 
   @override
-  double calculateConstraints(double availableWidth) =>
-      availableWidth * fraction;
+  double calculateConstraints(double availableWidth) => availableWidth * fraction;
 }
 
 /// Indicates that a column will take the remaining space in the viewport.
@@ -62,8 +59,7 @@ final class RemainingColumnSize extends ColumnSize {
   const RemainingColumnSize();
 
   @override
-  double calculateConstraints(double availableWidth) =>
-      math.max(0.0, availableWidth);
+  double calculateConstraints(double availableWidth) => math.max(0.0, availableWidth);
 }
 
 /// A column size that uses the maximum value of two provided constraints.
@@ -82,11 +78,9 @@ final class MaxColumnSize extends ColumnSize {
   int get hashCode => Object.hash(a.hashCode, b.hashCode);
 
   @override
-  bool operator ==(Object other) =>
-      other is MaxColumnSize && other.a == a && other.b == b;
+  bool operator ==(Object other) => other is MaxColumnSize && other.a == a && other.b == b;
 
   @override
-  double calculateConstraints(double availableWidth) => math.max(
-      a.calculateConstraints(availableWidth),
-      b.calculateConstraints(availableWidth));
+  double calculateConstraints(double availableWidth) =>
+      math.max(a.calculateConstraints(availableWidth), b.calculateConstraints(availableWidth));
 }
