@@ -52,7 +52,7 @@ class _DoubleListRowsState<K extends Comparable<K>, T> extends State<_DoubleList
       child: ScrollConfiguration(
         behavior: ScrollConfiguration.of(context).copyWith(scrollbars: false),
         child: Opacity(
-          opacity: widget.controller._state == TableState.idle ? 1 : 0.5,
+          opacity: widget.controller.isIdle() ? 1 : 0.5,
           child: Scrollbar(
             thumbVisibility: theme.verticalScrollbarVisibility,
             controller: normalController,
@@ -64,7 +64,7 @@ class _DoubleListRowsState<K extends Comparable<K>, T> extends State<_DoubleList
                     primary: false,
                     controller: fixedController,
                     itemCount: widget.controller._totalItems,
-                    separatorBuilder: (_, __) => const Divider(height: 0, color: Color(0xFFD6D6D6)),
+                    separatorBuilder: (_, _) => const Divider(height: 0, color: Color(0xFFD6D6D6)),
                     itemBuilder: (context, index) => _FixedPartRow<K, T>(
                       index: index,
                       fixedColumnCount: widget.fixedColumnCount,
@@ -83,11 +83,12 @@ class _DoubleListRowsState<K extends Comparable<K>, T> extends State<_DoubleList
                       children: [
                         ConstrainedBox(
                           constraints: BoxConstraints(
-                              maxWidth: widget.sizes.skip(widget.fixedColumnCount).fold(0.0, (a, b) => a + b)),
+                            maxWidth: widget.sizes.skip(widget.fixedColumnCount).fold(0.0, (a, b) => a + b),
+                          ),
                           child: ListView.separated(
                             controller: normalController,
                             itemCount: widget.controller._totalItems,
-                            separatorBuilder: (_, __) => const Divider(height: 0, color: Color(0xFFD6D6D6)),
+                            separatorBuilder: (_, _) => const Divider(height: 0, color: Color(0xFFD6D6D6)),
                             itemBuilder: (context, index) => _VariablePartRow<K, T>(
                               sizes: widget.sizes,
                               index: index,
