@@ -16,14 +16,11 @@ class DefaultFooter<K extends Comparable<K>, T> extends StatelessWidget {
       children: [
         if (child != null) Expanded(child: child!) else const Spacer(),
         RefreshTable<K, T>(),
-        const VerticalDivider(
-            color: Color(0xFFD6D6D6), width: 3, indent: 10, endIndent: 10),
+        const VerticalDivider(color: Color(0xFFD6D6D6), width: 3, indent: 10, endIndent: 10),
         PageSizeSelector<K, T>(),
-        const VerticalDivider(
-            color: Color(0xFFD6D6D6), width: 3, indent: 10, endIndent: 10),
+        const VerticalDivider(color: Color(0xFFD6D6D6), width: 3, indent: 10, endIndent: 10),
         CurrentPage<K, T>(),
-        const VerticalDivider(
-            color: Color(0xFFD6D6D6), width: 3, indent: 10, endIndent: 10),
+        const VerticalDivider(color: Color(0xFFD6D6D6), width: 3, indent: 10, endIndent: 10),
         NavigationButtons<K, T>(),
       ],
     );
