@@ -44,11 +44,8 @@ class LinkedScrollControllerGroup {
 
   /// Creates a new controller that is linked to any existing ones.
   ScrollController addAndGet() {
-    final initialScrollOffset = _attachedControllers.isEmpty
-        ? 0.0
-        : _attachedControllers.first.position.pixels;
-    final controller =
-        _LinkedScrollController(this, initialScrollOffset: initialScrollOffset);
+    final initialScrollOffset = _attachedControllers.isEmpty ? 0.0 : _attachedControllers.first.position.pixels;
+    final controller = _LinkedScrollController(this, initialScrollOffset: initialScrollOffset);
     _allControllers.add(controller);
     controller.addListener(_offsetNotifier.notifyListeners);
     return controller;
@@ -68,15 +65,10 @@ class LinkedScrollControllerGroup {
       _allControllers.where((controller) => controller.hasClients);
 
   /// Animates the scroll position of all linked controllers to [offset].
-  Future<void> animateTo(
-    double offset, {
-    required Curve curve,
-    required Duration duration,
-  }) async {
+  Future<void> animateTo(double offset, {required Curve curve, required Duration duration}) async {
     final animations = <Future<void>>[];
     for (final controller in _attachedControllers) {
-      animations
-          .add(controller.animateTo(offset, duration: duration, curve: curve));
+      animations.add(controller.animateTo(offset, duration: duration, curve: curve));
     }
     return Future.wait<void>(animations).then<void>((List<void> _) => null);
   }
@@ -124,9 +116,7 @@ class _LinkedScrollControllerGroupOffsetNotifier extends ChangeNotifier {
 class _LinkedScrollController extends ScrollController {
   final LinkedScrollControllerGroup _controllers;
 
-  _LinkedScrollController(this._controllers,
-      {required super.initialScrollOffset})
-      : super(keepScrollOffset: false);
+  _LinkedScrollController(this._controllers, {required super.initialScrollOffset}) : super(keepScrollOffset: false);
 
   @override
   void dispose() {
@@ -137,19 +127,21 @@ class _LinkedScrollController extends ScrollController {
   @override
   void attach(ScrollPosition position) {
     assert(
-        position is _LinkedScrollPosition,
-        '_LinkedScrollControllers can only be used with'
-        ' _LinkedScrollPositions.');
-    final _LinkedScrollPosition linkedPosition =
-        position as _LinkedScrollPosition;
-    assert(linkedPosition.owner == this,
-        '_LinkedScrollPosition cannot change controllers once created.');
+      position is _LinkedScrollPosition,
+      '_LinkedScrollControllers can only be used with'
+      ' _LinkedScrollPositions.',
+    );
+    final _LinkedScrollPosition linkedPosition = position as _LinkedScrollPosition;
+    assert(linkedPosition.owner == this, '_LinkedScrollPosition cannot change controllers once created.');
     super.attach(position);
   }
 
   @override
-  _LinkedScrollPosition createScrollPosition(ScrollPhysics physics,
-      ScrollContext context, ScrollPosition? oldPosition) {
+  _LinkedScrollPosition createScrollPosition(
+    ScrollPhysics physics,
+    ScrollContext context,
+    ScrollPosition? oldPosition,
+  ) {
     return _LinkedScrollPosition(
       this,
       physics: physics,
@@ -160,9 +152,8 @@ class _LinkedScrollController extends ScrollController {
   }
 
   @override
-  double get initialScrollOffset => _controllers._attachedControllers.isEmpty
-      ? super.initialScrollOffset
-      : _controllers.offset;
+  double get initialScrollOffset =>
+      _controllers._attachedControllers.isEmpty ? super.initialScrollOffset : _controllers.offset;
 
   @override
   _LinkedScrollPosition get position => super.position as _LinkedScrollPosition;
@@ -174,9 +165,7 @@ class _LinkedScrollController extends ScrollController {
 
   Iterable<_LinkedScrollActivity> linkWithPeers(_LinkedScrollPosition driver) {
     assert(canLinkWithPeers);
-    return _allPeersWithClients
-        .map((peer) => peer.link(driver))
-        .expand((e) => e);
+    return _allPeersWithClients.map((peer) => peer.link(driver)).expand((e) => e);
   }
 
   Iterable<_LinkedScrollActivity> link(_LinkedScrollPosition driver) {
@@ -243,9 +232,7 @@ class _LinkedScrollPosition extends ScrollPositionWithSingleContext {
     if (newPixels == pixels) {
       return 0.0;
     }
-    updateUserScrollDirection(newPixels - pixels > 0.0
-        ? ScrollDirection.forward
-        : ScrollDirection.reverse);
+    updateUserScrollDirection(newPixels - pixels > 0.0 ? ScrollDirection.forward : ScrollDirection.reverse);
 
     if (owner.canLinkWithPeers) {
       _peerActivities.addAll(owner.linkWithPeers(this));
@@ -266,9 +253,7 @@ class _LinkedScrollPosition extends ScrollPositionWithSingleContext {
     if (value == pixels) {
       return;
     }
-    updateUserScrollDirection(value - pixels > 0.0
-        ? ScrollDirection.forward
-        : ScrollDirection.reverse);
+    updateUserScrollDirection(value - pixels > 0.0 ? ScrollDirection.forward : ScrollDirection.reverse);
 
     if (owner.canLinkWithPeers) {
       _peerActivities.addAll(owner.linkWithPeers(this));
@@ -288,8 +273,7 @@ class _LinkedScrollPosition extends ScrollPositionWithSingleContext {
     if (this.activity is! _LinkedScrollActivity) {
       beginActivity(_LinkedScrollActivity(this));
     }
-    final _LinkedScrollActivity activity =
-        this.activity as _LinkedScrollActivity;
+    final _LinkedScrollActivity activity = this.activity as _LinkedScrollActivity;
     activity.link(driver);
     return activity;
   }
