@@ -8,12 +8,15 @@ abstract class _RowBuilder<K extends Comparable<K>, T> extends StatefulWidget {
   @override
   State<StatefulWidget> createState() => _RowBuilderState<K, T>();
 
-  List<Widget> buildColumns(BuildContext context, int index,
-      PagedDataTableController<K, T> controller, PagedDataTableThemeData theme);
+  List<Widget> buildColumns(
+    BuildContext context,
+    int index,
+    PagedDataTableController<K, T> controller,
+    PagedDataTableThemeData theme,
+  );
 }
 
-class _RowBuilderState<K extends Comparable<K>, T>
-    extends State<_RowBuilder<K, T>> {
+class _RowBuilderState<K extends Comparable<K>, T> extends State<_RowBuilder<K, T>> {
   late final controller = TableControllerProvider.of<K, T>(context);
   late final theme = PagedDataTableTheme.of(context);
   bool selected = false;
@@ -30,9 +33,7 @@ class _RowBuilderState<K extends Comparable<K>, T>
 
   @override
   Widget build(BuildContext context) {
-    Widget child = Row(
-        children:
-            widget.buildColumns(context, widget.index, controller, theme));
+    Widget child = Row(children: widget.buildColumns(context, widget.index, controller, theme));
     var color = theme.rowColor?.call(widget.index);
     if (selected && theme.selectedRow != null) {
       color = theme.selectedRow;
@@ -78,10 +79,11 @@ class _FixedPartRow<K extends Comparable<K>, T> extends _RowBuilder<K, T> {
 
   @override
   List<Widget> buildColumns(
-      BuildContext context,
-      int index,
-      PagedDataTableController<K, T> controller,
-      PagedDataTableThemeData theme) {
+    BuildContext context,
+    int index,
+    PagedDataTableController<K, T> controller,
+    PagedDataTableThemeData theme,
+  ) {
     final item = controller._currentDataset[index];
     final list = <Widget>[];
 
@@ -110,10 +112,11 @@ class _VariablePartRow<K extends Comparable<K>, T> extends _RowBuilder<K, T> {
 
   @override
   List<Widget> buildColumns(
-      BuildContext context,
-      int index,
-      PagedDataTableController<K, T> controller,
-      PagedDataTableThemeData theme) {
+    BuildContext context,
+    int index,
+    PagedDataTableController<K, T> controller,
+    PagedDataTableThemeData theme,
+  ) {
     final item = controller._currentDataset[index];
     final list = <Widget>[];
 
@@ -127,8 +130,14 @@ class _VariablePartRow<K extends Comparable<K>, T> extends _RowBuilder<K, T> {
   }
 }
 
-Widget _buildCell<T>(BuildContext context, int index, T value, double width,
-    PagedDataTableThemeData theme, ReadOnlyTableColumn column) {
+Widget _buildCell<T>(
+  BuildContext context,
+  int index,
+  T value,
+  double width,
+  PagedDataTableThemeData theme,
+  ReadOnlyTableColumn column,
+) {
   Widget child = Container(
     padding: theme.cellPadding,
     margin: theme.padding,
