@@ -1,12 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:paged_datatable/src/base_theme.dart';
 
-final class PagedDataTableThemeData {
-  /// The padding of the cell.
-  final EdgeInsetsGeometry cellPadding;
-
-  /// The padding of a entire column, from other columns.
-  final EdgeInsetsGeometry padding;
-
+final class PagedDataTableThemeData extends FilterBarThemeData {
   /// The [BorderRadius] of the table.
   final BorderRadius borderRadius;
 
@@ -24,9 +19,6 @@ final class PagedDataTableThemeData {
 
   /// The height of each row.
   final double rowHeight;
-
-  /// The filter bar's height.
-  final double filterBarHeight;
 
   /// The [BoxBorder] to render for each cell.
   final BoxBorder cellBorderSide;
@@ -52,56 +44,50 @@ final class PagedDataTableThemeData {
   /// A flag that indicates if the horizontal scrollbar should be visible.
   final bool horizontalScrollbarVisibility;
 
-  /// The width breakpoint that [PagedDataTable] uses to decide if will render a popup or a bottom sheet when the filter dialog is requested.
-  final double filterDialogBreakpoint;
-
-  /// The [ChipThemeData] to apply to filter chips.
-  final ChipThemeData? chipTheme;
-
   const PagedDataTableThemeData({
-    this.cellPadding =
-        const EdgeInsets.symmetric(horizontal: 8.0, vertical: 6.0),
-    this.padding = const EdgeInsets.symmetric(horizontal: 16.0),
+    super.cellPadding = const EdgeInsets.symmetric(horizontal: 8.0, vertical: 6.0),
+    super.padding = const EdgeInsets.symmetric(horizontal: 16.0),
+    super.filterBarHeight = 50.0,
+    super.filterDialogBreakpoint = 1000.0,
+    super.chipTheme,
     this.borderRadius = const BorderRadius.all(Radius.circular(4.0)),
     this.elevation = 0.0,
     this.cellBorderSide = const Border(),
     this.headerHeight = 56.0,
     this.footerHeight = 56.0,
-    this.filterBarHeight = 50.0,
     this.rowHeight = 52.0,
     this.selectedRow,
-    this.cellTextStyle =
-        const TextStyle(color: Colors.black, overflow: TextOverflow.ellipsis),
+    this.cellTextStyle = const TextStyle(color: Colors.black, overflow: TextOverflow.ellipsis),
     this.headerTextStyle = const TextStyle(
-        color: Colors.black,
-        fontWeight: FontWeight.bold,
-        overflow: TextOverflow.ellipsis),
+      color: Colors.black,
+      fontWeight: FontWeight.bold,
+      overflow: TextOverflow.ellipsis,
+    ),
     this.footerTextStyle = const TextStyle(fontSize: 14, color: Colors.black),
     this.rowColor,
     this.verticalScrollbarVisibility = true,
     this.horizontalScrollbarVisibility = true,
-    this.filterDialogBreakpoint = 1000.0,
-    this.chipTheme,
     this.backgroundColor = Colors.white,
   });
 
   @override
   int get hashCode => Object.hash(
-      cellPadding,
-      padding,
-      borderRadius,
-      elevation,
-      headerHeight,
-      footerHeight,
-      rowHeight,
-      cellBorderSide,
-      cellTextStyle,
-      headerTextStyle,
-      rowColor,
-      verticalScrollbarVisibility,
-      horizontalScrollbarVisibility,
-      chipTheme,
-      backgroundColor);
+    cellPadding,
+    padding,
+    borderRadius,
+    elevation,
+    headerHeight,
+    footerHeight,
+    rowHeight,
+    cellBorderSide,
+    cellTextStyle,
+    headerTextStyle,
+    rowColor,
+    verticalScrollbarVisibility,
+    horizontalScrollbarVisibility,
+    chipTheme,
+    backgroundColor,
+  );
 
   @override
   bool operator ==(Object other) =>
@@ -120,8 +106,7 @@ final class PagedDataTableThemeData {
           other.cellBorderSide == cellBorderSide &&
           other.selectedRow == selectedRow &&
           other.verticalScrollbarVisibility == verticalScrollbarVisibility &&
-          other.horizontalScrollbarVisibility ==
-              horizontalScrollbarVisibility &&
+          other.horizontalScrollbarVisibility == horizontalScrollbarVisibility &&
           other.chipTheme == chipTheme &&
           other.backgroundColor == backgroundColor);
 }
@@ -129,18 +114,15 @@ final class PagedDataTableThemeData {
 final class PagedDataTableTheme extends InheritedWidget {
   final PagedDataTableThemeData data;
 
-  const PagedDataTableTheme(
-      {required this.data, required super.child, super.key});
+  const PagedDataTableTheme({required this.data, required super.child, super.key});
 
   @override
-  bool updateShouldNotify(covariant InheritedWidget oldWidget) =>
-      data != (oldWidget as PagedDataTableTheme).data;
+  bool updateShouldNotify(covariant InheritedWidget oldWidget) => data != (oldWidget as PagedDataTableTheme).data;
 
   /// Lookups for a [PagedDataTableTheme] widget in the widget tree, if not found, returns null.
   static PagedDataTableThemeData? maybeOf(BuildContext context) =>
       context.dependOnInheritedWidgetOfExactType<PagedDataTableTheme>()?.data;
 
   /// Lookups for a [PagedDataTableTheme] widget in the widget tree, if not found, then default [PagedDataTableThemeData] is returned.
-  static PagedDataTableThemeData of(BuildContext context) =>
-      maybeOf(context) ?? const PagedDataTableThemeData();
+  static PagedDataTableThemeData of(BuildContext context) => maybeOf(context) ?? const PagedDataTableThemeData();
 }
